requirements.txt
streamlit~=1.37.1
plotly~=5.23.0
pandas~=2.2.2
openai~=1.41.1
PyYAML~=6.0.2
altair~=5.4.0